excercises 4.15 - 4.21

+++ controllers/blogs.js
@@ -3,16 +3,10 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-  return null
-}
 
 
 blogsRouter.get('/', async (request, response) => {
+  //excercise 4.17
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1})
     response.json(blogs)
   })
@@ -20,9 +14,11 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body  
 
-  const token = getTokenFrom(request)
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!token || !decodedToken.id) {
+  //excercise 4.18 - 4.19
+ 
+  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
   const user = await User.findById(decodedToken.id)
@@ -54,11 +50,26 @@ blogsRouter.post('/', async (request, response, next) => {
    
   })
   
-  //excercise 4.13
+  //excercise 4.13 & 4.21
   blogsRouter.delete('/:id', async (request, response, next) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
     
+
+    const blog = await Blog.findById(request.params.id)
+
+    if(blog.user.toString() === user._id.toString()){
       await Blog.findByIdAndRemove(request.params.id)
       response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'cannot delete blogs from other users' })
+    }
+   
+    
+      
     
   })
